--- src/components/Modals/LaunchQuestModal/CreateLaunchQuestModal.type.ts
@@ -1,12 +1,0 @@
-import { Team } from '@Domain/Team'
-
-interface CreateLaunchQuestProps {
-  isOpened: boolean
-  teams: Team[]
-}
-
-interface CreateLaunchQuestEmits {
-  (event: 'close-modal'): void
-}
-
-export { CreateLaunchQuestProps, CreateLaunchQuestEmits }
