--- src/components/Modals/LaunchQuestModal/LaunchQuestModal.type.ts
@@ -1,5 +1,0 @@
-interface LaunchQuestModalProps {
-  canGoBack?: boolean
-}
-
-export { LaunchQuestModalProps }
