--- src/components/Modals/LaunchQuestModal/PassLaunchQuestModal.type.ts
@@ -1,12 +1,0 @@
-import { LaunchQuest } from '@Domain/Quest'
-
-interface PassLaunchQuestProps {
-  isOpened: boolean
-  launchQuest: LaunchQuest | null
-}
-
-interface PassLaunchQuestEmits {
-  (event: 'close-modal'): void
-}
-
-export { PassLaunchQuestProps, PassLaunchQuestEmits }
