+++ src/components/Modals/QuestModal/QuestModal.vue
@@ -1,28 +1,38 @@
 <script lang="ts" setup>
-import { ref, onMounted } from 'vue'
+import { ref, onMounted, computed } from 'vue'
 import { storeToRefs } from 'pinia'
 import { useRouter, useRoute } from 'vue-router'
-import Select from '@Components/Inputs/Select/Select.vue'
 
 import Button from '@Components/Button/Button.vue'
+import Select from '@Components/Inputs/Select/Select.vue'
 import Typography from '@Components/Typography/Typography.vue'
-import { QuestModalProps } from '@Components/Modals/QuestModal/QuestModal.type'
+import { LaunchQuestModalProps } from '@Components/Modals/QuestModal/QuestModal.type'
 
 import ModalLayout from '@Layouts/ModalLayout/ModalLayout.vue'
 
 import useUserStore from '@Store/user/userStore'
-import useQuestsStore from '@Store/quests/questsStore'
+import useQuestTemplatesStore from '@Store/questTemplates/questTemplatesStore'
 import IndicatorItem from '@Components/IndicatorItem/IndicatorItem.vue'
+import useQuestsStore from '@Store/quests/questsStore'
+import { Quest } from '@Domain/Quest'
+import useTeamStore from '@Store/teams/teamsStore'
 import { OptionType } from '@Components/Inputs/Select/Select.types'
 import { useForm } from 'vee-validate'
 
-const props = defineProps<QuestModalProps>()
+defineProps<LaunchQuestModalProps>()
 
 const userStore = useUserStore()
 const { user } = storeToRefs(userStore)
 
-const questStore = useQuestsStore()
-const { quest } = storeToRefs(questStore)
+const questTemplatesStore = useQuestTemplatesStore()
+const { questTemplate } = storeToRefs(questTemplatesStore)
+
+const launchQuestsStore = useQuestsStore()
+const { quests: launchQuests } = storeToRefs(launchQuestsStore)
+const launchQuest = ref<Quest>()
+
+const teamsStore = useTeamStore()
+const { teams } = storeToRefs(teamsStore)
 
 const router = useRouter()
 const route = useRoute()
@@ -39,7 +49,7 @@ const changeAvailability = handleSubmit(async (model) => {
   if (currentUser?.token) {
     const { token } = currentUser
     const available = model.available
-    if (available == quest.value?.available) {
+    if (available == questTemplate.value?.available) {
       setFieldError('available', 'Значение должно отличаться от предыдущего')
     }
   }
@@ -47,12 +57,22 @@ const changeAvailability = handleSubmit(async (model) => {
 
 onMounted(async () => {
   const currentUser = user.value
-  const id = route.params.id.toString()
+  const idLaunchQuest = route.params.idLaunchQuest.toString()
 
   if (currentUser?.token) {
     const { token } = currentUser
-    await questStore.getQuest(id, token)
-    setValues({ available: quest.value?.available })
+    await launchQuestsStore.getQuests(token)
+    launchQuest.value = launchQuests.value.find(
+      (item) => item.idQuest == idLaunchQuest,
+    )
+    if (!launchQuest.value) return
+    await questTemplatesStore.getQuestTemplate(
+      launchQuest.value.idQuestTemplate,
+      token,
+    )
+    setValues({ available: launchQuest.value?.available })
+    const teamsId = launchQuest.value.idTeams
+    await teamsStore.getTeamsByIds(teamsId, token)
   }
 })
 
@@ -70,6 +90,14 @@ const availableOptions: OptionType[] = [
     label: 'Не доступен',
   },
 ]
+
+const computedRole = computed(() => {
+  return user.value?.role
+})
+
+const computedQuestAvailability = computed(() => {
+  return launchQuest.value?.available ? 'Открыт' : 'Скрыт'
+})
 </script>
 
 <template>
@@ -91,7 +119,7 @@ const availableOptions: OptionType[] = [
 
           <Typography
             class-name="p-2 w-100 bg-white rounded-3 fs-4 text-primary text-nowrap overflow-scroll-hidden"
-            >{{ quest?.name ? quest?.name : 'Опрос' }}
+            >{{ questTemplate?.name ? questTemplate?.name : 'Опрос' }}
           </Typography>
         </div>
 
@@ -99,11 +127,23 @@ const availableOptions: OptionType[] = [
           <div class="profile-modal__info mb-3">
             <div class="bg-white rounded-3 border p-3 gap-3 w-100">
               <div class="w-100 border-bottom pb-1">
+                <Typography class-name="fs-5 text-primary"
+                  >Дата начала и окончания:</Typography
+                >
+              </div>
+
+              <div class="d-flex flex-column gap-3 mt-3">
+                <p>{{ launchQuest?.startAt + ' - ' + launchQuest?.endAt }}</p>
+              </div>
+            </div>
+
+            <div class="bg-white rounded-3 border p-3 gap-3 w-100">
+              <div class="w-100 border-bottom pb-1">
                 <Typography class-name="fs-5 text-primary">Описание:</Typography>
               </div>
 
               <div class="d-flex flex-column gap-3 mt-3">
-                <p>{{ quest?.description }}</p>
+                <p>{{ questTemplate?.description }}</p>
               </div>
             </div>
           </div>
@@ -111,26 +151,67 @@ const availableOptions: OptionType[] = [
           <div class="profile-modal__info mb-3">
             <div class="bg-white rounded-3 border p-3 gap-3 w-100">
               <div class="w-100 border-bottom pb-1">
+                <Typography class-name="fs-5 text-primary"
+                  >Название шаблона опроса:</Typography
+                >
+              </div>
+
+              <div class="d-flex flex-column gap-3 mt-3">
+                <p>{{ questTemplate?.name }}</p>
+              </div>
+            </div>
+            <div class="bg-white rounded-3 border p-3 gap-3 w-100">
+              <div class="w-100 border-bottom pb-1">
                 <Typography class-name="fs-5 text-primary">Доступность:</Typography>
               </div>
 
               <div class="d-flex gap-3 mt-3">
-                <Select
-                  name="available"
-                  :options="availableOptions"
-                ></Select>
-
-                <Button
-                  @click="changeAvailability"
-                  variant="primary"
-                  class-name="w-fit"
-                  >Изменить</Button
+                <p v-if="computedRole != 'PROJECT_OFFICE'">
+                  {{ computedQuestAvailability }}
+                </p>
+                <div
+                  class="d-flex justify-content-between w-100 gap-3"
+                  v-else
                 >
+                  <Select
+                    class-name="w-100"
+                    name="available"
+                    :options="availableOptions"
+                  ></Select>
+
+                  <Button
+                    @click="changeAvailability"
+                    variant="primary"
+                    class-name="w-fit"
+                    >Изменить</Button
+                  >
+                </div>
               </div>
             </div>
           </div>
         </div>
-        <div class="bg-white rounded-3 border p-3 gap-3 w-100">
+        <div
+          v-if="computedRole == 'PROJECT_OFFICE'"
+          class="bg-white rounded-3 border p-3 gap-3 w-100"
+        >
+          <div class="w-100 border-bottom pb-1">
+            <Typography class-name="fs-5 text-primary">Список команд:</Typography>
+          </div>
+
+          <div class="d-flex flex-wrap gap-3 mt-3 justify-content-between">
+            <div
+              class="w-49"
+              v-for="team in teams"
+              :key="team.id"
+            >
+              <p>{{ team.name }}</p>
+            </div>
+          </div>
+        </div>
+        <div
+          v-if="computedRole == 'PROJECT_OFFICE'"
+          class="bg-white rounded-3 border p-3 gap-3 w-100 mt-3"
+        >
           <div class="w-100 border-bottom pb-1">
             <Typography class-name="fs-5 text-primary">Список вопросов:</Typography>
           </div>
@@ -138,7 +219,7 @@ const availableOptions: OptionType[] = [
           <div class="d-flex flex-wrap gap-3 mt-3 justify-content-between">
             <div
               class="w-49"
-              v-for="indicator in quest?.indicators"
+              v-for="indicator in questTemplate?.indicators"
               :key="indicator.idIndicator"
             >
               <IndicatorItem :indicator="indicator" />
