+++ src/components/Modals/QuestTemplateModal/CreateQuestTemplateModal.type.ts
@@ -1,0 +1,10 @@
+interface CreateQuestModalProps {
+  isOpened: boolean
+  id: string | null
+}
+
+interface CreateQuestModalEmits {
+  (event: 'close-modal'): void
+}
+
+export { CreateQuestModalProps, CreateQuestModalEmits }
