+++ src/components/Modals/QuestTemplateModal/QuestTemplateModal.type.ts
@@ -1,0 +1,5 @@
+interface QuestTemplateModalProps {
+  canGoBack?: boolean
+}
+
+export { QuestTemplateModalProps }
