+++ src/components/Modals/QuestTemplateModal/QuestTemplateModal.vue
@@ -1,38 +1,28 @@
 <script lang="ts" setup>
-import { ref, onMounted, computed } from 'vue'
+import { ref, onMounted } from 'vue'
 import { storeToRefs } from 'pinia'
 import { useRouter, useRoute } from 'vue-router'
+import Select from '@Components/Inputs/Select/Select.vue'
 
 import Button from '@Components/Button/Button.vue'
-import Select from '@Components/Inputs/Select/Select.vue'
 import Typography from '@Components/Typography/Typography.vue'
-import { LaunchQuestModalProps } from '@Components/Modals/LaunchQuestModal/LaunchQuestModal.type'
+import { QuestTemplateModalProps } from '@Components/Modals/QuestTemplateModal/QuestTemplateModal.type'
 
 import ModalLayout from '@Layouts/ModalLayout/ModalLayout.vue'
 
 import useUserStore from '@Store/user/userStore'
-import useQuestsStore from '@Store/quests/questsStore'
+import useQuestTemplatesStore from '@Store/questTemplates/questTemplatesStore'
 import IndicatorItem from '@Components/IndicatorItem/IndicatorItem.vue'
-import useLaunchQuestStore from '@Store/launchQuests/launchQuestsStore'
-import { LaunchQuest } from '@Domain/Quest'
-import useTeamStore from '@Store/teams/teamsStore'
 import { OptionType } from '@Components/Inputs/Select/Select.types'
 import { useForm } from 'vee-validate'
 
-defineProps<LaunchQuestModalProps>()
+const props = defineProps<QuestTemplateModalProps>()
 
 const userStore = useUserStore()
 const { user } = storeToRefs(userStore)
 
-const questsStore = useQuestsStore()
-const { quest } = storeToRefs(questsStore)
-
-const launchQuestsStore = useLaunchQuestStore()
-const { launchQuests } = storeToRefs(launchQuestsStore)
-const launchQuest = ref<LaunchQuest>()
-
-const teamsStore = useTeamStore()
-const { teams } = storeToRefs(teamsStore)
+const questTemplatesStore = useQuestTemplatesStore()
+const { questTemplate: quest } = storeToRefs(questTemplatesStore)
 
 const router = useRouter()
 const route = useRoute()
@@ -57,18 +47,12 @@ const changeAvailability = handleSubmit(async (model) => {
 
 onMounted(async () => {
   const currentUser = user.value
-  const idLaunchQuest = route.params.idLaunchQuest.toString()
+  const id = route.params.id.toString()
 
   if (currentUser?.token) {
     const { token } = currentUser
-    await launchQuestsStore.getLaunchQuests(token)
-    launchQuest.value = launchQuests.value.find(
-      (item) => item.idLaunchQuest == idLaunchQuest,
-    )
-    if (!launchQuest.value) return
-    await questsStore.getQuest(launchQuest.value.id, token)
-    setValues({ available: launchQuest.value?.available })
-    await teamsStore.getTeamsByIds(launchQuest.value.idTeams, token)
+    await questTemplatesStore.getQuestTemplate(id, token)
+    setValues({ available: quest.value?.available })
   }
 })
 
@@ -86,14 +70,6 @@ const availableOptions: OptionType[] = [
     label: 'Не доступен',
   },
 ]
-
-const computedRole = computed(() => {
-  return user.value?.role
-})
-
-const computedQuestAvailability = computed(() => {
-  return launchQuest.value?.available ? 'Открыт' : 'Скрыт'
-})
 </script>
 
 <template>
@@ -123,18 +99,6 @@ const computedQuestAvailability = computed(() => {
           <div class="profile-modal__info mb-3">
             <div class="bg-white rounded-3 border p-3 gap-3 w-100">
               <div class="w-100 border-bottom pb-1">
-                <Typography class-name="fs-5 text-primary"
-                  >Дата начала и окончания:</Typography
-                >
-              </div>
-
-              <div class="d-flex flex-column gap-3 mt-3">
-                <p>{{ launchQuest?.startAt + ' - ' + launchQuest?.endAt }}</p>
-              </div>
-            </div>
-
-            <div class="bg-white rounded-3 border p-3 gap-3 w-100">
-              <div class="w-100 border-bottom pb-1">
                 <Typography class-name="fs-5 text-primary">Описание:</Typography>
               </div>
 
@@ -147,67 +111,26 @@ const computedQuestAvailability = computed(() => {
           <div class="profile-modal__info mb-3">
             <div class="bg-white rounded-3 border p-3 gap-3 w-100">
               <div class="w-100 border-bottom pb-1">
-                <Typography class-name="fs-5 text-primary"
-                  >Название шаблона опроса:</Typography
-                >
-              </div>
-
-              <div class="d-flex flex-column gap-3 mt-3">
-                <p>{{ quest?.name }}</p>
-              </div>
-            </div>
-            <div class="bg-white rounded-3 border p-3 gap-3 w-100">
-              <div class="w-100 border-bottom pb-1">
                 <Typography class-name="fs-5 text-primary">Доступность:</Typography>
               </div>
 
               <div class="d-flex gap-3 mt-3">
-                <p v-if="computedRole != 'PROJECT_OFFICE'">
-                  {{ computedQuestAvailability }}
-                </p>
-                <div
-                  class="d-flex justify-content-between w-100 gap-3"
-                  v-else
-                >
-                  <Select
-                    class-name="w-100"
-                    name="available"
-                    :options="availableOptions"
-                  ></Select>
+                <Select
+                  name="available"
+                  :options="availableOptions"
+                ></Select>
 
-                  <Button
-                    @click="changeAvailability"
-                    variant="primary"
-                    class-name="w-fit"
-                    >Изменить</Button
-                  >
-                </div>
+                <Button
+                  @click="changeAvailability"
+                  variant="primary"
+                  class-name="w-fit"
+                  >Изменить</Button
+                >
               </div>
             </div>
           </div>
         </div>
-        <div
-          v-if="computedRole == 'PROJECT_OFFICE'"
-          class="bg-white rounded-3 border p-3 gap-3 w-100"
-        >
-          <div class="w-100 border-bottom pb-1">
-            <Typography class-name="fs-5 text-primary">Список команд:</Typography>
-          </div>
-
-          <div class="d-flex flex-wrap gap-3 mt-3 justify-content-between">
-            <div
-              class="w-49"
-              v-for="team in teams"
-              :key="team.id"
-            >
-              <p>{{ team.name }}</p>
-            </div>
-          </div>
-        </div>
-        <div
-          v-if="computedRole == 'PROJECT_OFFICE'"
-          class="bg-white rounded-3 border p-3 gap-3 w-100 mt-3"
-        >
+        <div class="bg-white rounded-3 border p-3 gap-3 w-100">
           <div class="w-100 border-bottom pb-1">
             <Typography class-name="fs-5 text-primary">Список вопросов:</Typography>
           </div>
