+++ src/components/Tables/QuestsTable/QuestTableCollapse.types.ts
