--- src/store/launchQuests/initialState.ts
@@ -1,7 +1,0 @@
-import { LaunchQuest } from '@Domain/Quest'
-
-interface InitialState {
-  launchQuests: LaunchQuest[]
-}
-
-export default InitialState
