+++ src/store/questTemplates/initialState.ts
@@ -1,0 +1,8 @@
+import { QuestTemplate, QuestTemplateShort } from '@Domain/Quest'
+
+interface InitialState {
+  questTemplates: QuestTemplateShort[]
+  questTemplate: QuestTemplate | null
+}
+
+export default InitialState
